+++ src/Repositories/AbstractModelRepository.php
@@ -95,6 +95,19 @@ abstract class AbstractModelRepository implements ModelRepositoryInterface
         return $models;
     }
 
+    /**
+     * @param string  $column
+     * @param mixed[] $values
+     * @return Collection<int, T>
+     */
+    public function whereNotIn(string $column, array $values): Collection
+    {
+        /** @var Collection<int, T> $models */
+        $models = $this->model->whereNotIn($column, $values)->get();
+
+        return $models;
+    }
+
     public function all(string ...$column): Collection
     {
         /** @var Collection<int, T> $models */