+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This changelog is initialized in release 1.0.0
 
 ### Added
 * WhereIn method to repository
+* WhereNotIn method to repository
 
 ## [v2.4.0] - 2022-06-03